create wrapper component to avoid redundant logic
