+++ polymorphia-frontend/components/equipment/modals/ChestModal.tsx
@@ -1,25 +1,19 @@
-import Modal from "@/components/modal/Modal";
 import { API_STATIC_URL } from "@/services/api";
 import XPCard from "@/components/xp-card/XPCard";
 import { Item } from "@/interfaces/equipment/EquipmentInterfaces";
-import { useContext, useEffect, useState } from "react";
+import { useContext } from "react";
 import { EquipmentContext } from "@/components/providers/EquipmentContext";
+import Modal from "@/components/modal/Modal";
 
 export default function ChestModal() {
   const { currentChestModalData, setCurrentChestModalData } =
     useContext(EquipmentContext);
   const chest = currentChestModalData;
-  const [modalVisible, setModalVisible] = useState(false);
-
-  useEffect(() => {
-    setModalVisible(!!chest);
-  }, [chest]);
 
   return (
     <Modal
-      isOpen={modalVisible}
-      onRequestClose={() => setModalVisible(false)}
-      onClosed={() => setCurrentChestModalData(null)}
+      isDataPresented={chest !== null}
+      onClose={() => setCurrentChestModalData(null)}
       title={chest?.title ?? ""}
       subtitle={chest?.subtitle ?? ""}
     >
