+++ polymorphia-frontend/components/equipment/modals/ItemModal.tsx
@@ -1,26 +1,20 @@
-import Modal from "@/components/modal/Modal";
 import "../../../styles/points-summary.css";
 import { API_STATIC_URL } from "@/services/api";
 import XPCard from "@/components/xp-card/XPCard";
 import { Item } from "@/interfaces/equipment/EquipmentInterfaces";
 import { EquipmentContext } from "@/components/providers/EquipmentContext";
-import { useContext, useEffect, useState } from "react";
+import { useContext } from "react";
+import Modal from "@/components/modal/Modal";
 
 export default function ItemModal() {
   const { currentItemModalData, setCurrentItemModalData } =
     useContext(EquipmentContext);
   const item = currentItemModalData;
-  const [modalVisible, setModalVisible] = useState(false);
-
-  useEffect(() => {
-    setModalVisible(!!item);
-  }, [item]);
 
   return (
     <Modal
-      isOpen={modalVisible}
-      onRequestClose={() => setModalVisible(false)}
-      onClosed={() => setCurrentItemModalData(null)}
+      isDataPresented={item !== null}
+      onClose={() => setCurrentItemModalData(null)}
       title={item?.title ?? ""}
       subtitle={item?.subtitle ?? ""}
     >
