+++ polymorphia-frontend/interfaces/course/event-section/EventSectionInterfaces.ts
@@ -171,7 +171,7 @@ export interface RewardsInfoProps {
 }
 
 export interface TestDetailsModalProps
-  extends Omit<ModalProps, "title" | "isOpen" | "children" | "onRequestClose"> {
+  extends Omit<ModalProps, "title" | "isDataPresented" | "children"> {
   eventSectionId: number;
   selectedGradableEventId: number | null;
 }
